create new folder and move App to components folder
